requirements.txt
# # dependency
# fastapi==0.115.7
# mysql-connector-python==9.2.0
# python-dotenv==1.0.1
# requests==2.32.3
# uvicorn==0.34.0

# dependency
fastapi
mysql-connector-python
python-dotenv
requests
uvicorn